chore: 상수 에러 메시지 libs로 분리
